+++ Debug/Types/Addon.lua
@@ -8,14 +8,24 @@
 ---@class WK_Core : AceAddon
 
 ---@class WK_Main
+---@field TableData WK_Main_TableData
+---@field TableColumns WK_Main_TableColumns
 
 ---@class WK_Checklist
+---@field TableData WK_Checklist_TableData
+---@field TableColumns WK_Checklist_TableColumns
 
 ---@class WK_Helpers
 
 ---@class WK_Constants
+---@field prefix string
+---@field commands string[]
 
 ---@class WK_Addon
+---@field name string
+---@field title string
+---@field version string
+---@field notes string
 ---@field Core WK_Core
 ---@field libs WK_Libs
 ---@field LiqUI LiqUI_Instance